Fixed #26401 -- Allow auth machinery without installing auth app
Refactored imports, and moved AnonymousUser to base_user so the auth
middleware, context processor and AnonymousUser can be used without
putting django.contrib.auth in INSTALLED_APPS

Part 1 of 2
