+++ django/contrib/auth/backends.py
@@ -1,7 +1,6 @@
 from __future__ import unicode_literals
 
 from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Permission
 
 
 class ModelBackend(object):
@@ -35,6 +34,7 @@ class ModelBackend(object):
         return user_obj.user_permissions.all()
 
     def _get_group_permissions(self, user_obj):
+        from django.contrib.auth.models import Permission
         user_groups_field = get_user_model()._meta.get_field('groups')
         user_groups_query = 'group__%s' % user_groups_field.related_query_name()
         return Permission.objects.filter(**{user_groups_query: user_obj})
@@ -45,6 +45,7 @@ class ModelBackend(object):
         be either "group" or "user" to return permissions from
         `_get_group_permissions` or `_get_user_permissions` respectively.
         """
+        from django.contrib.auth.models import Permission
         if not user_obj.is_active or user_obj.is_anonymous or obj is not None:
             return set()
